+++ py/desiutil/install.py
@@ -173,10 +173,10 @@ class DesiInstall(object):
     """
     cross_install_host = 'edison'
     nersc_hosts = ('cori', 'edison', 'datatran', 'scigate')
-    default_nersc_dir_templates = {'edison': '/global/common/edison/contrib/desi/desiconda/{desiconda_version}',
-                                   'cori': '/global/common/cori/contrib/desi/desiconda/{desiconda_version}',
-                                   'datatran': '/global/project/projectdirs/desi/software/datatran/desiconda/{desiconda_version}',
-                                   'scigate': '/global/project/projectdirs/desi/software/scigate/desiconda/{desiconda_version}'}
+    default_nersc_dir_templates = {'edison': '/global/common/software/desi/edison{knl}/desiconda/{desiconda_version}',
+                                   'cori': '/global/common/software/desi/cori{knl}/desiconda/{desiconda_version}',
+                                   'datatran': '/global/common/software/desi/datatran/desiconda/{desiconda_version}',
+                                   'scigate': '/global/common/software/desi/scigate/desiconda/{desiconda_version}'}
 
     def __init__(self, test=False):
         """Bare-bones initialization.
@@ -254,6 +254,8 @@ class DesiInstall(object):
                             dest='force',
                             help=('Overwrite any existing installation of ' +
                                   'this product/version.'))
+        parser.add_argument('-K', '--knl', action='store_true', dest='knl',
+                            help='Support KNL versions of desiconda (e.g. coriknl).')
         parser.add_argument('-k', '--keep', action='store_true',
                             dest='keep',
                             help='Keep the exported build directory.')
@@ -617,6 +619,12 @@ class DesiInstall(object):
                     build_type.add('src')
         return build_type
 
+    @property
+    def knl(self):
+        """String for use in specifying the name of KNL-based installs.
+        """
+        return ('', 'knl')[int(self.options.knl)]
+
     def anaconda_version(self):
         """Try to determine the exact DESI+Anaconda version from the
         environment.
@@ -649,8 +657,8 @@ class DesiInstall(object):
             Path to the host-specific install directory.
         """
         if nersc_host is None:
-            return self.default_nersc_dir_templates[self.nersc].format(desiconda_version=self.options.anaconda)
-        return self.default_nersc_dir_templates[nersc_host].format(desiconda_version=self.options.anaconda)
+            return self.default_nersc_dir_templates[self.nersc].format(knl=self.knl, desiconda_version=self.options.anaconda)
+        return self.default_nersc_dir_templates[nersc_host].format(knl=self.knl, desiconda_version=self.options.anaconda)
 
     def set_install_dir(self):
         """Decide on an install directory.
@@ -754,7 +762,7 @@ class DesiInstall(object):
             return None
         else:
             if self.baseproduct == 'desimodules':
-                nersc_module = join(self.default_nersc_dir_templates[self.nersc].format(desiconda_version='startup'),
+                nersc_module = join(self.default_nersc_dir_templates[self.nersc].format(knl=self.knl, desiconda_version='startup'),
                                     'modulefiles')
             else:
                 nersc_module = join(self.default_nersc_dir(),
@@ -882,11 +890,10 @@ class DesiInstall(object):
                 out, err = proc.communicate()
                 status = proc.returncode
                 log.debug(out)
-                # Temporarily ignore all error messages from script.
-                # if status != 0 and len(err) > 0:
-                #     message = "Error grabbing extra data: {0}".format(err)
-                #     log.critical(message)
-                #     raise DesiInstallException(message)
+                if status != 0 and len(err) > 0:
+                    message = "Error grabbing extra data: {0}".format(err)
+                    log.critical(message)
+                    raise DesiInstallException(message)
         return
 
     def copy_install(self):